ssize_t is not standard

+++ include/ccf/crypto/cose.h
@@ -2,7 +2,6 @@
 // Licensed under the Apache 2.0 License.
 #pragma once
 
-#include <cstddef>
 #include <cstdint>
 #include <span>
 #include <variant>
@@ -17,7 +16,7 @@ namespace ccf::cose::edit
 
     struct AtKey
     {
-      ssize_t key;
+      int64_t key;
     };
 
     using Type = std::variant<InArray, AtKey>;
@@ -42,7 +41,7 @@ namespace ccf::cose::edit
    */
   std::vector<uint8_t> set_unprotected_header(
     const std::span<const uint8_t>& cose_input,
-    ssize_t key,
+    int64_t key,
     pos::Type position,
     const std::vector<uint8_t> value);
 }
